+++ src/main/java/com/supermartijn642/trashcans/screen/ItemTrashCanScreen.java
@@ -5,14 +5,14 @@ import com.supermartijn642.core.gui.GuiGraphicsHelper;
 import com.supermartijn642.trashcans.TrashCanBlockEntity;
 import com.supermartijn642.trashcans.TrashCans;
 import com.supermartijn642.trashcans.packet.PacketToggleItemWhitelist;
-import net.minecraft.resources.ResourceLocation;
+import net.minecraft.resources.Identifier;
 
 /**
  * Created 7/11/2020 by SuperMartijn642
  */
 public class ItemTrashCanScreen extends TrashCanScreen<ItemTrashCanContainer> {
 
-    public static final ResourceLocation BACKGROUND = ResourceLocation.fromNamespaceAndPath("trashcans", "item_screen");
+    public static final Identifier BACKGROUND = Identifier.fromNamespaceAndPath("trashcans", "item_screen");
 
     private WhitelistButton whitelistButton;
 
@@ -33,7 +33,7 @@ public class ItemTrashCanScreen extends TrashCanScreen<ItemTrashCanContainer> {
     }
 
     @Override
-    protected ResourceLocation getBackground(){
+    protected Identifier getBackground(){
         return BACKGROUND;
     }