+++ changelog.md
@@ -1,2 +1,2 @@
 ### Trash Cans 1.0.18
-- Initial release of Trash Cans for Minecraft 1.21.9 & 1.21.10
+- Initial release of Trash Cans for Minecraft 1.21.11